Add my second job
If Ministry of Ideas was an agency Local Global used to be a start-up.

Yes, the very first start-up I joined :tada:

Local Global worked on a web portal which was a thing then.

We were building a set of web applications that provided the following
services:

- Webmail
- Video-sharing service
- File sharing and storage service

A company was also maintaining a news website which got attacked due to
political reasons during presidential elections.

I had to fight those attacks which made me really curious about
information security, so I started to look deeper into that...