Describe my journey into security
So my previous job at Local Global gave me a hands-on experience with
cyber-attacks and security which fascinated me.

I decided to go deeper into that industry and joined information
security team at [one of the largest nation-wide banks][1] in Ukraine.

As team grew into a standalone department I was promoted to lead one
of the teams inside that department that essentially focused on
_how_ the bank should tackle information security problems as well as
researching new and existing products and processes from information
security standpoint.

[1]: https://en.wikipedia.org/wiki/First_Ukrainian_International_Bank