Add job at Kreditech
My job at Kreditech was my first job in Germany.

Kreditech was a fintech start-up. At its core was an idea of scoring
underbanked people using their digital footprint on the internet and
giving out loans based on that score.

I was interviewed for a security position following my work at a bank,
but end up as a DevOps engineer because that skillset was more needed
to support launch of newly acquired company's flagship product Monedo
Pay for which I completely set up the whole infrastructure from the
ground up.

While working at Kreditech I got exposed to Node.js.

I saw huge potential in that technology and wanted to focus more on it
by making it my full-time job.